watchmate/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
